requirements.txt
Django==1.6.2
Markdown==2.3.1
PyYAML==3.10
coverage==3.7.1
django-filter==0.7
django-markupfield==1.2.0
django-model-utils==2.0
django-suit==0.2.5
djangorestframework==2.3.12
geoip2==0.4.2
ipaddr==2.1.11
maxminddb==0.2.1
psycopg2==2.5.2
python-dateutil==2.2
requests==2.2.1
six==1.5.2
wsgiref==0.1.2
